rename core -> server -> dcx for readability and semantic
