+++ src/dcx/index.ts
@@ -1,6 +1,6 @@
 export * from './web5-manager.js';
 export * from './config.js';
-export * from './dcx-server.js';
+export * from './server.js';
 export * from './handlers.js';
-import server from './dcx-server.js';
+import server from './server.js';
 export default server;
