+++ src/index.ts
@@ -1,6 +1,6 @@
-export * from './server/index.js';
+export * from './dcx/index.js';
 export * from './protocol/index.js';
 export type * from './types/index.js';
 export * from './utils/index.js';
-import server from './server/index.js';
+import server from './dcx/index.js';
 export default server;
